创建ship类
